map step to section and handle next click
