+++ frontend/webapp/containers/setup/destination/destination.section.tsx
@@ -1,0 +1,15 @@
+import React, { useEffect, useMemo, useState } from "react";
+
+export function DestinationSection({ sectionData, setSectionData }: any) {
+  //   const { isLoading, data } = useQuery([QUERIES.API_NAMESPACES], getNamespaces);
+
+  //   if (isLoading) {
+  //     return null;
+  //   }
+
+  return (
+    <>
+      <div>DestinationSection</div>
+    </>
+  );
+}
