+++ frontend/webapp/containers/setup/setup.section/setup.section.tsx
@@ -1,152 +1,91 @@
-import React, { useEffect, useMemo, useState } from "react";
+import React, { useMemo, useState } from "react";
 import {
   SetupContentWrapper,
   SetupSectionContainer,
+  StepListWrapper,
 } from "./setup.section.styled";
 import { SetupHeader } from "../setup.header/setup.header";
-import { SourcesList, SourcesOptionMenu } from "@/components/setup";
-import { getApplication, setNamespaces } from "@/services/setup";
+import { setNamespaces } from "@/services/setup";
+import { SETUP } from "@/utils/constants";
+import Steps from "@/design.system/steps/steps";
+import { SourcesSection } from "../sources/sources.section";
+import { DestinationSection } from "../destination/destination.section";
 
-const DEFAULT_CONFIG = {
-  selected_all: false,
-  future_selected: false,
-};
+const STEPS = [
+  {
+    index: 1,
+    id: "choose-source",
+    title: SETUP.STEPS.CHOOSE_SOURCE,
+    status: SETUP.STEPS.STATUS.ACTIVE,
+  },
+  {
+    index: 2,
+    id: "choose-destination",
+    title: SETUP.STEPS.CHOOSE_DESTINATION,
+    status: SETUP.STEPS.STATUS.DISABLED,
+  },
+  {
+    index: 3,
+    id: "create-connection",
+    title: SETUP.STEPS.CREATE_CONNECTION,
+    status: SETUP.STEPS.STATUS.DISABLED,
+  },
+];
 
-export function SetupSection({ namespaces }: any) {
-  const [currentNamespace, setCurrentNamespace] = useState<any>(null);
-  const [selectedApplications, setSelectedApplications] = useState<any>({});
-  const [searchFilter, setSearchFilter] = useState<string>("");
-
-  const namespacesList = useMemo(() => {
-    return namespaces.map((item: any, index: number) => {
-      return { id: index, label: item.name };
-    });
-  }, [namespaces]);
-
-  useEffect(() => {
-    !currentNamespace && setCurrentNamespace(namespaces[0]);
-  }, [namespaces]);
-
-  useEffect(() => {
-    onNameSpaceChange();
-  }, [currentNamespace]);
+export function SetupSection() {
+  const [sectionData, setSectionData] = useState<any>({});
+  const [steps, setSteps] = useState(STEPS);
+  const [currentStep, setCurrentStep] = useState(STEPS[0]);
 
   const totalSelected = useMemo(() => {
     let total = 0;
-    for (const key in selectedApplications) {
-      const apps = selectedApplications[key]?.objects;
+    for (const key in sectionData) {
+      const apps = sectionData[key]?.objects;
       const counter = apps?.filter((item: any) => item.selected)?.length;
       total += counter;
     }
     return total;
-  }, [JSON.stringify(selectedApplications)]);
+  }, [JSON.stringify(sectionData)]);
 
-  const sourceData = useMemo(() => {
-    const data = selectedApplications[currentNamespace?.name];
-    return searchFilter
-      ? data?.objects.filter((item: any) =>
-          item.name.toLowerCase().includes(searchFilter.toLowerCase())
-        )
-      : data?.objects;
-  }, [searchFilter, currentNamespace, selectedApplications]);
-
-  async function onNameSpaceChange() {
-    if (!currentNamespace || selectedApplications[currentNamespace?.name])
-      return;
-    const data = await getApplication(currentNamespace?.name);
-    const newSelectedNamespace = {
-      ...selectedApplications,
-      [currentNamespace?.name]: {
-        ...DEFAULT_CONFIG,
-        objects: [...data?.applications],
-      },
-    };
-
-    setSelectedApplications(newSelectedNamespace);
-  }
-
-  function handleSourceClick({ item }: any) {
-    const objIndex = selectedApplications[
-      currentNamespace?.name
-    ].objects.findIndex((app) => app.name === item.name);
-
-    const namespace = selectedApplications[currentNamespace?.name];
-    let objects = [...namespace.objects];
-
-    objects[objIndex].selected = !objects[objIndex].selected;
-
-    let currentNamespaceConfig = {
-      ...namespace,
-      objects,
-    };
-
-    if (!objects[objIndex].selected && namespace.selected_all) {
-      currentNamespaceConfig = {
-        ...currentNamespaceConfig,
-        selected_all: false,
-        future_selected: false,
-      };
+  function renderCurrentSection() {
+    switch (currentStep?.id) {
+      case "choose-source":
+        return (
+          <SourcesSection
+            sectionData={sectionData}
+            setSectionData={setSectionData}
+          />
+        );
+      case "choose-destination":
+        return <DestinationSection />;
+      default:
+        return null;
     }
-    handleSetNewSelectedConfig(currentNamespaceConfig);
   }
 
-  function onSelectAllChange(value: boolean) {
-    const namespace = selectedApplications[currentNamespace?.name];
-    let objects = [...namespace.objects];
-    objects.forEach((item) => {
-      item.selected = value;
-    });
-
-    const currentNamespaceConfig = {
-      future_selected: value,
-      selected_all: value,
-      objects,
-    };
-    handleSetNewSelectedConfig(currentNamespaceConfig);
-  }
-
-  function onFutureApplyChange(value: boolean) {
-    const currentNamespaceConfig = {
-      ...selectedApplications[currentNamespace?.name],
-      future_selected: value,
-    };
-    handleSetNewSelectedConfig(currentNamespaceConfig);
-  }
-
-  function handleSetNewSelectedConfig(config: any) {
-    const newSelectedNamespaceConfig = {
-      ...selectedApplications,
-      [currentNamespace?.name]: config,
-    };
-    setSelectedApplications(newSelectedNamespaceConfig);
+  function handleNamespacesUpdate() {
+    // setNamespaces(sectionData);
+    setCurrentStep(STEPS[1]);
   }
 
   function onNextClick() {
-    setNamespaces(selectedApplications);
+    switch (currentStep?.id) {
+      case "choose-source":
+        handleNamespacesUpdate();
+      default:
+        return null;
+    }
   }
 
   return (
-    <SetupSectionContainer>
-      <SetupHeader onNextClick={onNextClick} totalSelected={totalSelected} />
-      <SetupContentWrapper>
-        <SourcesOptionMenu
-          currentNamespace={currentNamespace}
-          setCurrentItem={setCurrentNamespace}
-          data={namespacesList}
-          searchFilter={searchFilter}
-          setSearchFilter={setSearchFilter}
-          onSelectAllChange={onSelectAllChange}
-          selectedApplications={selectedApplications}
-          onFutureApplyChange={onFutureApplyChange}
-        />
-
-        <SourcesList
-          data={sourceData}
-          selectedData={selectedApplications[currentNamespace?.name]}
-          onItemClick={handleSourceClick}
-          onClearClick={() => onSelectAllChange(false)}
-        />
-      </SetupContentWrapper>
-    </SetupSectionContainer>
+    <>
+      <StepListWrapper>
+        <Steps data={steps} />
+      </StepListWrapper>
+      <SetupSectionContainer>
+        <SetupHeader onNextClick={onNextClick} totalSelected={totalSelected} />
+        <SetupContentWrapper>{renderCurrentSection()}</SetupContentWrapper>
+      </SetupSectionContainer>
+    </>
   );
 }
